Revert "[Fixed] Run CI on a single machine by generating cloudbuild files (#4186)"
This reverts commit 7bd157e2c28f9c1a56134a78d674c86886dfc7b7.
