+++ .gitignore
@@ -59,4 +59,3 @@ tfjs-backend-wasm/wasm-out/*.js
 tfjs-backend-wasm/wasm-out/*.wasm
 yalc.lock
 yarn-error.log
-cloudbuild_generated.yml
