+++ e2e/cloudbuild.yml
@@ -21,8 +21,8 @@ steps:
   id: 'test'
   args: ['test-ci']
   env: ['BROWSERSTACK_USERNAME=deeplearnjs1', 'NIGHTLY=$_NIGHTLY']
+
   secretEnv: ['BROWSERSTACK_KEY']
-  waitFor: ['yarn', 'build-deps']
 
 secrets:
 - kmsKeyName: projects/learnjs-174218/locations/global/keyRings/tfjs/cryptoKeys/enc
