--- scripts/cloudbuild.yml
@@ -1,14 +1,0 @@
-steps:
-# Install common dependencies.
-- name: 'node:10'
-  id: 'yarn-common'
-  entrypoint: 'yarn'
-  args: ['install']
-
-# Test generate_cloudbuild.js
-- name: 'node:10'
-  dir: 'scripts'
-  id: 'test-generate-cloudbuild'
-  entrypoint: 'yarn'
-  args: ['test-generate-cloudbuild']
-  waitFor: ['yarn-common']
