+++ scripts/package_dependencies.json
@@ -1,18 +1,13 @@
+
 {
   "tfjs-core": [],
-  "tfjs-converter": ["tfjs-core", "tfjs-backend-cpu"],
+  "tfjs-converter": ["tfjs-core"],
   "tfjs-layers": ["tfjs-core", "tfjs-backend-cpu", "tfjs-backend-webgl"],
-  "tfjs-data": ["tfjs-backend-cpu", "tfjs-core", "tfjs-layers"],
+  "tfjs-data": ["tfjs-core", "tfjs-layers"],
   "tfjs": ["tfjs-core", "tfjs-converter", "tfjs-layers", "tfjs-data", "tfjs-backend-cpu", "tfjs-backend-webgl"],
   "tfjs-node": ["tfjs", "tfjs-core"],
   "tfjs-node-gpu": ["tfjs", "tfjs-core"],
   "tfjs-backend-wasm": ["tfjs-core", "tfjs-backend-cpu"],
   "tfjs-backend-cpu": ["tfjs-core"],
-  "tfjs-backend-webgl": ["tfjs-core", "tfjs-backend-cpu"],
-  "tfjs-backend-webgpu": [],
-  "tfjs-inference": [],
-  "tfjs-react-native": [],
-  "tfjs-vis": [],
-  "scripts": [],
-  "e2e": ["tfjs", "tfjs-backend-cpu", "tfjs-backend-webgl", "tfjs-converter", "tfjs-core", "tfjs-data", "tfjs-layers", "tfjs-node", "tfjs-backend-wasm"]
+  "tfjs-backend-webgl": ["tfjs-core", "tfjs-backend-cpu"]
 }
