+++ tfjs-backend-cpu/cloudbuild.yml
@@ -1,4 +1,3 @@
-
 steps:
 # Install common dependencies.
 - name: 'node:10'
