+++ tfjs-node-gpu/cloudbuild.yml
@@ -19,7 +19,7 @@ steps:
   entrypoint: 'yarn'
   id: 'yarn'
   args: ['install']
-  waitFor: ['yarn-common', 'prep-gpu']
+  waitFor: ['yarn-common']
 
 # Build add-on.
 - name: 'node:10'
@@ -35,7 +35,7 @@ steps:
   entrypoint: 'yarn'
   id: 'build-deps'
   args: ['build-deps-ci']
-  waitFor: ['yarn-common', 'prep-gpu']
+  waitFor: ['yarn-common']
 
 # Build.
 - name: 'node:10'
@@ -67,7 +67,7 @@ steps:
   entrypoint: 'yarn'
   id: 'ensure-cpu-gpu-packages-align'
   args: ['ensure-cpu-gpu-packages-align']
-  waitFor: ['yarn-common', 'prep-gpu']
+  waitFor: ['yarn-common']
 
 timeout: 1800s
 logsBucket: 'gs://tfjs-build-logs'
