+++ tfjs-vis/cloudbuild.yml
@@ -1,4 +1,3 @@
-
 steps:
 # Install common dependencies.
 - name: 'node:10'
@@ -18,7 +17,6 @@ steps:
 - name: 'node:10'
   dir: 'tfjs-vis'
   entrypoint: 'yarn'
-  id: 'test-vis'
   args: ['test-ci']
   env: ['BROWSERSTACK_USERNAME=deeplearnjs1']
   secretEnv: ['BROWSERSTACK_KEY']
